Refactor CosmosDB folder structure
  - This is in preparation for adding new CosmosDB resources.
