+++ pkg/resourcemanager/cosmosdb/account/cosmosdb_reconcile.go
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-package cosmosdbs
+package account
 
 import (
 	"context"
